Moving eventtype components
