+++ components/pages/eventtypes/CreateFirstEventTypeView.tsx
@@ -3,7 +3,7 @@ import { inferQueryOutput } from "@lib/trpc";
 
 import UserCalendarIllustration from "@components/ui/svg/UserCalendarIllustration";
 
-import CreateNewEventDialog from "./CreateNewEventDialog";
+import CreateNewEventButton from "./CreateNewEventButton";
 
 type Profiles = inferQueryOutput<"viewer.eventTypes">["profiles"];
 
@@ -21,7 +21,7 @@ const CreateFirstEventTypeView = ({ canAddEvents, profiles }: Props) => {
       <div className="block mx-auto text-center md:max-w-screen-sm">
         <h3 className="mt-2 text-xl font-bold text-neutral-900">{t("new_event_type_heading")}</h3>
         <p className="mt-1 mb-2 text-md text-neutral-600">{t("new_event_type_description")}</p>
-        <CreateNewEventDialog canAddEvents={canAddEvents} profiles={profiles} />
+        <CreateNewEventButton canAddEvents={canAddEvents} profiles={profiles} />
       </div>
     </div>
   );
